@@ -1,39 +1,29 @@
 [tool.poetry]
 name = "vn-open-api-provinces"
-version = "0.3.0"
+version = "0.4.0"
 description = ""
 authors = ["Nguyễn Hồng Quân <ng.hong.quan@gmail.com>"]
 license = "GPL-3.0-or-later"
 
 [tool.poetry.dependencies]
-python = "^3.9"
-fastapi = "^0.95.1"
-pydantic = "^1.9.1"
-Logbook = "^1.5.3"
-uvicorn = "^0.22.0"
-fastapi-rfc7807 = "^0.5.0"
-aiofiles = "^23.1.0"
-lunr = "^0.6.2"
-Unidecode = "^1.3.6"
+python = "^3.12"
+fastapi = "^0.111.0"
+pydantic = "^2.8.2"
+Logbook = "^1.7.0.post0"
+uvicorn = "^0.30.1"
+aiofiles = "^24.1.0"
+lunr = "^0.7.0.post1"
+Unidecode = "^1.3.8"
 single-version = "^1.5.1"
 vietnam-provinces = "^0.5.0"
+pydantic-settings = "^2.3.4"
 
 [tool.poetry.group.dev.dependencies]
-doc8 = "^1.1.1"
-wheel = "^0.40.0"
-pytest = "^7.3.1"
-
-[tool.black]
-line-length = 120
-target-version = ['py310']
-include = '\.pyi?$'
-skip-string-normalization = true
+pytest = "^8.3.2"
+ruff = "^0.5.1"
 
 [tool.ruff]
 line-length = 120
-# Enable Pyflakes `E` and `F` codes by default.
-select = ["E", "F"]
-ignore = []
 # Exclude a variety of commonly ignored directories.
 exclude = [
   ".bzr",
@@ -56,10 +46,17 @@ exclude = [
   "node_modules",
   "venv",
 ]
+target-version = "py312"
+
+[tool.ruff.lint]
+# Enable Pyflakes `E` and `F` codes by default.
+select = ["E", "F"]
+ignore = []
 # Allow unused variables when underscore-prefixed.
 dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
-# Assume Python 3.10.
-target-version = "py310"
+
+[tool.ruff.format]
+quote-style = "single"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
