@@ -1,7 +1,7 @@
 from enum import Enum
-from typing import List, Dict, Tuple
+from typing import Annotated
 
-from pydantic import BaseModel, Field
+from pydantic import BaseModel, Field, ConfigDict
 
 from vietnam_provinces.base import VietNamDivisionType
 
@@ -23,21 +23,17 @@ _EXAMPLE_WARD = {
     'code': 1,
     'division_type': 'phường',
     'codename': 'phuong_phuc_xa',
-    'district_code': 1
+    'district_code': 1,
 }
 
 
 class Ward(BaseModel):
+    model_config = ConfigDict(json_schema_extra={'examples': [_EXAMPLE_WARD]})
     name: str
     code: int
     division_type: VietNamDivisionType
     codename: str
     district_code: int
-
-    class Config:
-        schema_extra = {
-            'example': _EXAMPLE_WARD
-        }
 
 
 _EXAMPLE_DISTRICT = {
@@ -46,22 +42,18 @@ _EXAMPLE_DISTRICT = {
     'division_type': 'quận',
     'codename': 'quan_ba_dinh',
     'province_code': 1,
-    'wards': [_EXAMPLE_WARD]
+    'wards': [_EXAMPLE_WARD],
 }
 
 
 class District(BaseModel):
+    model_config = ConfigDict(json_schema_extra={'examples': [_EXAMPLE_DISTRICT]})
     name: str
     code: int
     division_type: VietNamDivisionType
     codename: str
     province_code: int
-    wards: List[Ward] = Field(default=[])
-
-    class Config:
-        schema_extra = {
-            'example': _EXAMPLE_DISTRICT
-        }
+    wards: Annotated[list[Ward], Field(default_factory=list)]
 
 
 _EXAMPLE_PROVINCE = {
@@ -70,42 +62,40 @@ _EXAMPLE_PROVINCE = {
     'division_type': 'thành phố trung ương',
     'codename': 'thanh_pho_ha_noi',
     'phone_code': 24,
-    'districts': [_EXAMPLE_DISTRICT]
+    'districts': [_EXAMPLE_DISTRICT],
 }
 
 
 class ProvinceResponse(BaseModel):
+    model_config = ConfigDict(json_schema_extra={'examples': [_EXAMPLE_PROVINCE]})
     name: str
     code: int
     division_type: VietNamDivisionType
     codename: str
     phone_code: int
-    districts: List[District] = Field(default=[])
-
-    class Config:
-        schema_extra = {
-            'example': _EXAMPLE_PROVINCE
-        }
+    districts: Annotated[list[District], Field(default_factory=list)]
 
 
 class VersionResponse(BaseModel):
     data_version: str
 
 
+_EXAMPLE_MATCH = {
+    'name': 'Thị xã Phú Mỹ',
+    'code': 754,
+    'matches': {'mỹ': (11, 13)},
+}
+
+
 class SearchResult(BaseModel):
+    model_config = ConfigDict(json_schema_extra={'examples': [_EXAMPLE_MATCH]})
     name: str
     code: int
-    matches: Dict[str, Tuple[int, int]] = Field({}, title='Matched words and their positions in name.',
-                                                description='This info can help client side highlight '
-                                                'the result in display.')
-
-    class Config:
-        schema_extra = {
-            'example': {
-                'name': 'Thị xã Phú Mỹ',
-                'code': 754,
-                'matches': {
-                    'mỹ': [11, 13]
-                },
-            }
-        }
+    matches: Annotated[
+        dict[str, tuple[int, int]],
+        Field(
+            {},
+            title='Matched words and their positions in name.',
+            description='This info can help client side highlight the result in display.',
+        ),
+    ]
